+++ src/HomePageView/HomePageView.jsx
@@ -1,6 +1,5 @@
-import { useParams } from 'react-router-dom'
+import { useParams, Link } from 'react-router-dom'
 import "./HomePageView.css"
-import SearchComponent from '../SearchComponent/SearchComponent.jsx'
 
 function HomePageView() {
   const displayName = useParams().username
@@ -8,7 +7,9 @@ function HomePageView() {
   return( 
     <div>
       <h1>{displayName}</h1>
-      <SearchComponent />
+      <Link to={'/pokemon_search'}>
+        <button>Pokemon</button>
+      </Link>
     </div>
   )   
 }
