+++ src/SearchComponent/SearchComponent.jsx
@@ -1,26 +1,27 @@
-import { useState, useEffect } from 'react'
-import { Link } from "react-router-dom"
-import "./SearchComponent.css"
-import CardsContainer from '../CardsContainer/CardsContainer'
+// import { useState, useEffect } from 'react'
+// import { Link } from "react-router-dom"
+// import "./SearchComponent.css"
+// import CardsContainer from '../CardsContainer/CardsContainer'
 
-const fetchPokemonCards = 'https://api.pokemontcg.io/v2/cards?page=1&pageSize=100'
-const fetchMTGCards = 'https://api.magicthegathering.io/v1/cards'
+// const fetchPokemonCards = 'https://api.pokemontcg.io/v2/cards?page=1&pageSize=100'
+// const fetchMTGCards = 'https://api.magicthegathering.io/v1/cards'
 
-function SearchComponent() {
-    const [cards, setCards] = useState([])
+// function SearchComponent() {
+    // const [cards, setCards] = useState([])
+    // const [showChild, setShowchild] = useState(false)
 
-    function getPokemonCards() {
-        fetch(fetchPokemonCards, {
-            headers: {
-                'Content-Type': 'application/json'
-            }
-        })
-        .then(response => response.json())
-        .then(data => {
-            setCards(data.data)
-        })
-        .catch(error => console.log('message: ', error.message))
-    }
+    // function getPokemonCards() {
+    //     fetch(fetchPokemonCards, {
+    //         headers: {
+    //             'Content-Type': 'application/json'
+    //         }
+    //     })
+    //     .then(response => response.json())
+    //     .then(data => {
+    //         setCards(data.data)
+    //     })
+    //     .catch(error => console.log('message: ', error.message))
+    // }
 
     // function getMTGCards() {
     //     fetch(fetchMTGCards, {
@@ -36,24 +37,22 @@ function SearchComponent() {
     //     .catch(error => console.log('message: ', error.message))
     // }
 
-    useEffect(() => {
-        getPokemonCards();
-    }, [])
+    // useEffect(() => {
+    //     getPokemonCards();
+    // }, [])
 
     // useEffect(() => {
     //     getMTGCards();
     // }, [])
 
-
-    return (
-        <div>
-            <Link to={'/pokemon_search'}>
+    // return (
+    //     <div>
+            {/* <Link to={'/pokemon_search'}>
                 <button onClick={() => getPokemonCards()}>Pokemon</button>
-            </Link>
-            <CardsContainer cards={ cards } />
+            </Link> */}
             {/* <button onClick={() => getMTGCards()}>MTG</button> */}
-        </div>
-    )
-}
+        {/* </div> */}
+    {/* )
+} */}
 
-export default SearchComponent
+// export default SearchComponent
